% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genReads.R
\name{genReads}
\alias{genReads}
\title{A function to generate reads with given error and save in fastq format}
\usage{
genReads(sequence, readLen = 10000, error = 0.1, coverage = 30)
}
\arguments{
\item{sequence}{sequence to generate reads from}

\item{readLen}{length of reads required}

\item{error}{percentage error in reads}

\item{coverage}{sequencing coverage desired}
}
\description{
A function to generate reads with given error and save in fastq format
}
